add colorZone

+++ Source/Charts/Renderers/YAxisRenderer.swift
@@ -96,14 +96,14 @@ open class YAxisRenderer: NSObject, AxisRenderer
         {
             let l = colorZones[i]
             var startPosition = CGPoint(x: 0.0, y: 0.0)
-            startPosition.x = l.start
+            startPosition.x = CGFloat(l.start)
             print(startPosition.x)
             startPosition.y = CGFloat(l.bottom)
             startPosition = startPosition.applying(trans)
         
             var endPosition = CGPoint(x: 0.0, y: 0.0)
             endPosition.y = CGFloat(l.top)
-            endPosition.x = l.end
+            endPosition.x = CGFloat(l.end)
             endPosition = endPosition.applying(trans)
             
             let rect = CGRect(x: min(startPosition.x, endPosition.x),
